Fix issue with urls not being correct #244
