+++ src/state/search/reducer.js
@@ -15,21 +15,15 @@ const initialState = {
 
 export default (state = initialState, action) => {
   switch (action.type) {
-    case 'SEARCH_RESULTS_FETCH': {
-      const { searchTerm } = action.data;
-      return {
-        ...state,
-        searchTerm: searchTerm,
-        isSearching: true
-      };
-    }
     case 'SEARCH_RESULTS_FETCH_SUCCESS': {
       const {
+        searchTerm,
         results,
         filters,
         totalResults,
         currentPage,
-        appliedFilters
+        appliedFilters,
+        resultsPerPage,
       } = action.data;
 
       const totalPages = Math.ceil(
@@ -38,22 +32,17 @@ export default (state = initialState, action) => {
 
       return {
         ...state,
+        searchTerm,        
         results,
         filters,
-        appliedFilters,
+        appliedFilters,        
         pagination: {
           ...state.pagination,
           totalResults,
           totalPages,
-          currentPage
+          currentPage,
+          resultsPerPage,
         },
-        isSearching: false
-      };
-    }
-    case 'SEARCH_RESULTS_FETCH_ERROR': {
-      return {
-        ...state,
-        isSearching: false
       };
     }
     case 'SEARCH_ORGANISMS_FETCH_SUCCESS': {
@@ -61,42 +50,6 @@ export default (state = initialState, action) => {
       return {
         ...state,
         organisms
-      };
-    }
-    case 'SEARCH_FILTER_TOGGLE': {
-      const { filterType, filterValue } = action.data;
-      const { appliedFilters: prevFilters } = state;
-      const appliedFilterType = new Set();
-      if (!prevFilters[filterType]) {
-        appliedFilterType.add(filterValue);
-      } else {
-        for (let filter of prevFilters[filterType]) {
-          if (filter !== filterValue) {
-            appliedFilterType.add(filter);
-          }
-        }
-        if (prevFilters[filterType].has(filterValue)) {
-          appliedFilterType.delete(filterValue);
-        } else {
-          appliedFilterType.add(filterValue);
-        }
-      }
-      return {
-        ...state,
-        appliedFilters: {
-          ...state.appliedFilters,
-          [filterType]: appliedFilterType
-        }
-      };
-    }
-    case 'UPDATE_PAGE_SIZE': {
-      const resultsPerPage = action.data;
-      return {
-        ...state,
-        pagination: {
-          ...state.pagination,
-          resultsPerPage
-        }
       };
     }
     default:
